making queries statement

+++ vmProjectBackend/Controllers/CourseController.cs
@@ -27,18 +27,30 @@ namespace vmProjectBackend.Controllers
 
         // GET: api/Course
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Course>>> GetCourses()
+        public async Task<ActionResult> GetCourses()
         {
             // grabbing the user that signed in
             string useremail = HttpContext.User.Identity.Name;
 
             // check if it is a professor
-            var user_prof = _context.Users.Where(p => p.email == useremail && p.userType == "Professor").FirstOrDefault();
+            var user_prof = _context.Users
+                            .Where(p => p.email == useremail && p.userType == "Professor")
+                            .FirstOrDefault();
 
-            Console.WriteLine("this is the user email" + useremail);
+            // Console.WriteLine("this is the user email" + useremail);
             if (user_prof != null)
             {
-                var listOfCourse = await _context.Enrollments.Where(u => u.UserId == user_prof.UserID).ToListAsync();
+                // var listOfCourse = await _context.Enrollments
+                // .Include(c => c.Course).Where(u => u.UserId == user_prof.UserID).Select(st => new
+                // {
+                //     name = st.Course.CourseName
+
+                // });
+                var listOfCourse = await _context.Enrollments
+                                .Include(c => c.Course)
+                                .Where(u => u.UserId
+                                            == user_prof.UserID)
+                                .ToListAsync();
                 return Ok(listOfCourse);
             }
 
@@ -49,19 +61,28 @@ namespace vmProjectBackend.Controllers
         }
 
         // GET: api/Course/5
-        [HttpGet("{id}", Name = "GetCourse")]
-        public async Task<ActionResult<Course>> GetCourse(long id)
+        [HttpGet("{id}/{section_num}/{semester}", Name = "GetCourse")]
+        public async Task<ActionResult<Course>> GetCourse(long id, string section_num, string semester)
         {
-            var course = await _context.Courses.FindAsync(id);
+            string useremail = HttpContext.User.Identity.Name;
+            // check if it is a professor
+            var user_prof = _context.Users
+                            .Where(p => p.email == useremail && p.userType == "Professor")
+                            .FirstOrDefault();
 
-            if (course == null)
+            if (user_prof != null)
             {
-                return NotFound();
+                var singleCourse = await _context.Enrollments
+                                .Include(c => c.Course)
+                                .Where(c => c.CourseID == id && c.section_num == section_num && c.semester == semester && c.UserId == user_prof.UserID)
+                                .ToListAsync();
+                return Ok(singleCourse);
             }
-
-            return course;
+            else
+            {
+                return NotFound("No such Course found");
+            }
         }
-
         // PUT: api/Course/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
@@ -98,40 +119,69 @@ namespace vmProjectBackend.Controllers
         [HttpPost]
         public async Task<ActionResult<Course>> PostCourse(Course course)
         {
-            _context.Courses.Add(course);
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateException)
-            {
-                if (CourseExists(course.CourseID))
-                {
-                    return Conflict();
-                }
-                else
-                {
-                    throw;
-                }
-            }
+            string useremail = HttpContext.User.Identity.Name;
+            // check if it is a professor
+            var user_prof = _context.Users
+                            .Where(p => p.email == useremail && p.userType == "Professor")
+                            .FirstOrDefault();
 
-            return CreatedAtAction("GetCourse", new { id = course.CourseID }, course);
+            if (user_prof != null)
+            {
+                _context.Courses.Add(course);
+                try
+                {
+                    await _context.SaveChangesAsync();
+                }
+                catch (DbUpdateException)
+                {
+                    if (CourseExists(course.CourseID))
+                    {
+                        return Conflict();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+
+                return CreatedAtAction("GetCourse", new { id = course.CourseID }, course);
+            }
+            return Unauthorized("You are not a professor");
+
         }
 
         // DELETE: api/Course/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteCourse(long id)
         {
-            var course = await _context.Courses.FindAsync(id);
-            if (course == null)
+            string useremail = HttpContext.User.Identity.Name;
+            // check if it is a professor
+            var user_prof = _context.Users
+                            .Where(p => p.email == useremail && p.userType == "Professor")
+                            .FirstOrDefault();
+
+            if (user_prof != null)
             {
-                return NotFound();
+                var singleCourse = _context.Enrollments
+                                               .Include(c => c.Course)
+                                               .Where(c => c.CourseID == id && c.UserId == user_prof.UserID)
+                                               .FirstOrDefault();
+                var course = await _context.Courses.FindAsync(id);
+                if (course == null)
+                {
+                    return NotFound();
+                }
+
+                _context.Courses.Remove(course);
+                await _context.SaveChangesAsync();
+
+                return NoContent();
             }
+            return Unauthorized("You are not a professor");
 
-            _context.Courses.Remove(course);
-            await _context.SaveChangesAsync();
 
-            return NoContent();
+            /***************************************************/
+
         }
 
         private bool CourseExists(long id)
@@ -139,15 +189,6 @@ namespace vmProjectBackend.Controllers
             return _context.Courses.Any(e => e.CourseID == id);
         }
 
-        // [HttpGet("coursenroll/{id}")]
-        // public async Task<IActionResult> GetCourseDetails(int id)
-        // {
-        //     var courseDetail = await _context.Courses
-        //     .Include()
-
-
-        //     return Ok();
-        // }
         [HttpPatch("{id}")]
         public async Task<ActionResult> PartialUpdate(long id, JsonPatchDocument<Course> patchDoc)
         {
